requirements.txt
flask
transformers
Pillow
requests
torch
